Sneaky bug: when you nest a multiple converter, the parent converter turns off exception throwing for the child converter This happens because the parent might want to check other converters, not hit an exception Anyway, turns out, once you have a repository that is not the main EAIResourceRepository, the structure manager wraps a localized multiple converter around the main one available from the factory This wrapping disables the exception throwing in the factory converter!
So this happens once you use a non-standard repository, for example for container artifacts (REST services!) so basically, once you have at least 1 rest service on your server, the conversion will no longer throw an exception.
If you don't have any rest services or other special things, the conversion will throw an error.

Turns out a lot of code is already depending on the dont-throw-an-exception logic, for example label evaluation in the vm service package.
So made that the standard, because we were getting non-working label evaluation on servers without REST services!